backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.115.14
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.22
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.46.2
typing_extensions==4.15.0
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
